promocode_service/requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
kafka-python
